refactored dependencies for GUI
