+++ Model.py
@@ -1,0 +1,18 @@
+class GameState:
+    def __init__(self, dealer_id):
+        self.dealer_id = dealer_id
+        self.team_tricks = [0, 0]
+        self.trump = None
+        self.alone = None
+        self.making_team = None
+
+class BidDecision:
+    def __init__(self, bid, alone):
+        self.bid = bid
+        self.alone = alone
+
+class SecondBidDecision:
+    def __init__(self, selected, trump, alone):
+        self.selected = selected
+        self.trump = trump
+        self.alone = alone
